Refactor sate and utils
- removeing Pause state
- make the formatTime reusable
