+++ mountain-guide-fe/app/(tabs)/active-hike.tsx
@@ -1,0 +1,1 @@
+
